# api_rest/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto DAW Backend API",
        "description": "API REST del Proyecto DAW. Esta API gestiona la lógica de negocio para la plataforma de intercambio de tiempo.\n *     \n *     Módulos incluidos:\n *     - **Autenticación**: Registro, login, logout y gestión de perfil.\n *     - **Usuarios**: Gestión completa de usuarios (CRUD).\n *     - **Servicios**: Gestión de ofertas y demandas de servicios.\n *     - **Transacciones**: Control de intercambios de tiempo entre usuarios.\n *     - **Valoraciones**: Sistema de reputación y feedback.\n *     - **Mensajes**: Sistema de mensajería interna.\n *     - **Common**: Recursos compartidos (provincias, poblaciones, etc.).",
        "contact": {
            "email": "soporte@proyectodaw.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo local"
        },
        {
            "url": "https://proyecto-daw-backend.onrender.com/api",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/mensajes": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener todos los mensajes",
                "operationId": "1c3bd7e1d8e74e4f2cb2e58e94fd8cc2",
                "responses": {
                    "200": {
                        "description": "Lista de mensajes obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Todos los mensajes obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/mensajes/{usuario_id}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener mensajes de un usuario (como emisor o receptor)",
                "operationId": "a039ed4b103d9ee4e952249b187eef6d",
                "parameters": [
                    {
                        "name": "usuario_id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensajes del usuario obtenidos correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensajes del usuario obtenidos correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 406
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/mensajes/{mensaje_id}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener mensaje por ID",
                "operationId": "3ad74695849beec83c4003d02418c116",
                "parameters": [
                    {
                        "name": "mensaje_id",
                        "in": "path",
                        "description": "ID del mensaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "time": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje obtenido correctamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 406
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/mensaje": {
            "post": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Crear un nuevo mensaje",
                "operationId": "9b6f2f6aa8e3c03f7404f7ed8a92a861",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emisor_id",
                                    "receptor_id",
                                    "servicio_id",
                                    "mensaje"
                                ],
                                "properties": {
                                    "emisor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "receptor_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "servicio_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "example": "Hola, estoy interesado en tu servicio"
                                    },
                                    "leido": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensaje creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/mensaje/{mensaje}": {
            "put": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Actualizar un mensaje existente",
                "operationId": "b5e90240036e758a4d33154aff793b4b",
                "parameters": [
                    {
                        "name": "mensaje",
                        "in": "path",
                        "description": "ID del mensaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "emisor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "receptor_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "servicio_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "example": "Mensaje actualizado"
                                    },
                                    "leido": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mensaje actualizado correctamente"
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Eliminar un mensaje",
                "operationId": "ca8623bf12813959742a2d01802e35e0",
                "parameters": [
                    {
                        "name": "mensaje",
                        "in": "path",
                        "description": "ID del mensaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje eliminado correctamente"
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticación mediante token Laravel Sanctum. Ingrese el token directamente.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}